[Theme] Reorganize translator ecosystem
