+++ src/Sylius/Bundle/ThemeBundle/Translation/Provider/Resource/CompositeTranslatorResourceProvider.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\ThemeBundle\Translation\Provider;
+namespace Sylius\Bundle\ThemeBundle\Translation\Provider\Resource;
 
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
