+++ src/Sylius/Bundle/ThemeBundle/Translation/Resource/TranslationResourceInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\ThemeBundle\Translation\Provider;
+namespace Sylius\Bundle\ThemeBundle\Translation\Resource;
 
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
