+++ src/Sylius/Bundle/ThemeBundle/spec/Translation/Provider/Resource/CompositeTranslatorResourceProviderSpec.php
@@ -12,10 +12,9 @@
 namespace spec\Sylius\Bundle\ThemeBundle\Translation\Provider;
 
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
-use Sylius\Bundle\ThemeBundle\Translation\Provider\CompositeTranslatorResourceProvider;
-use Sylius\Bundle\ThemeBundle\Translation\Provider\TranslationResourceInterface;
-use Sylius\Bundle\ThemeBundle\Translation\Provider\TranslatorResourceProviderInterface;
+use Sylius\Bundle\ThemeBundle\Translation\Provider\Resource\CompositeTranslatorResourceProvider;
+use Sylius\Bundle\ThemeBundle\Translation\Provider\Resource\TranslatorResourceProviderInterface;
+use Sylius\Bundle\ThemeBundle\Translation\Resource\TranslationResourceInterface;
 
 /**
  * @mixin CompositeTranslatorResourceProvider
@@ -26,7 +25,7 @@ final class CompositeTranslatorResourceProviderSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\ThemeBundle\Translation\Provider\CompositeTranslatorResourceProvider');
+        $this->shouldHaveType(CompositeTranslatorResourceProvider::class);
     }
 
     function it_implements_translator_resource_provider_interface()
