+++ src/Sylius/Bundle/ThemeBundle/spec/Translation/Resource/ThemeTranslationResourceSpec.php
@@ -12,25 +12,27 @@
 namespace spec\Sylius\Bundle\ThemeBundle\Translation\Provider;
 
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
-use Sylius\Bundle\ThemeBundle\Translation\Provider\TranslationResource;
-use Sylius\Bundle\ThemeBundle\Translation\Provider\TranslationResourceInterface;
+use Sylius\Bundle\ThemeBundle\Model\ThemeInterface;
+use Sylius\Bundle\ThemeBundle\Translation\Resource\ThemeTranslationResource;
+use Sylius\Bundle\ThemeBundle\Translation\Resource\TranslationResourceInterface;
 
 /**
- * @mixin TranslationResource
+ * @mixin ThemeTranslationResource
  *
  * @author Kamil Kokot <kamil.kokot@lakion.com>
  */
-final class TranslationResourceSpec extends ObjectBehavior
+final class ThemeTranslationResourceSpec extends ObjectBehavior
 {
-    function let()
+    function let(ThemeInterface $theme)
     {
-        $this->beConstructedWith('my-domain.my-locale.my-format');
+        $theme->getName()->willReturn('theme/name');
+
+        $this->beConstructedWith($theme, 'my-domain.my-locale.my-format');
     }
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\ThemeBundle\Translation\Provider\TranslationResource');
+        $this->shouldHaveType(ThemeTranslationResource::class);
     }
 
     function it_implements_translation_resource_interface()
@@ -42,13 +44,15 @@ final class TranslationResourceSpec extends ObjectBehavior
     {
         $this->getName()->shouldReturn('my-domain.my-locale.my-format');
         $this->getDomain()->shouldReturn('my-domain');
-        $this->getLocale()->shouldReturn('my-locale');
+        $this->getLocale()->shouldReturn('my-locale@theme-name');
         $this->getFormat()->shouldReturn('my-format');
     }
 
-    function it_throws_an_invalid_argument_exception_if_failed_to_instantiate_with_given_filepath()
+    function it_throws_an_invalid_argument_exception_if_failed_to_instantiate_with_given_filepath(ThemeInterface $theme)
     {
-        $this->beConstructedWith('one.dot');
+        $theme->getName()->willReturn('theme/name');
+
+        $this->beConstructedWith($theme, 'one.dot');
 
         $this->shouldThrow(\InvalidArgumentException::class)->duringInstantiation();
     }
